chore: remove the legacy --env-migration / --env-migration-dry-run commands
Fully remove the very old legacy config-migration commands and everything exclusive to them (multi-agent mapped):
- flags: EnvMigration, EnvMigrationDry, and the --old-env / LegacyEnvPath they alone consumed (internal/cli/args.go)
- mode handlers + dispatch + exclusivity entries (main_config_modes.go, main_modes.go)
- command impl cmd/proxsave/env_migration.go (+ test) - deleted
- internal/config/migration.go (EnvMigrationSummary, MigrateLegacyEnv, PlanLegacyEnvMigration + all helpers) (+ test)
  - deleted; these had ZERO non-env-migration callers (verified)
- help text (main_footer.go, install.go, config_helpers.go) + the config.go comment
- ALL doc citations: deleted docs/MIGRATION_GUIDE.md + docs/BACKUP_ENV_MAPPING.md; scrubbed README/DEVELOPER_GUIDE/
  EXAMPLES/TROUBLESHOOTING/INSTALL/CLI_REFERENCE (sections, table rows, links, TOC entries)

KEPT (shared, still used elsewhere): parseEnvFile, blockValueKeys, BackupCephConfig parsing, resolveInstallConfigPath,
and the captureStdout test helper (extracted to cmd/proxsave/testhelpers_test.go, used by 6 other test files). Built
via a multi-agent map -> remove -> 2-refuter workflow. build/vet + cmd/cli/config tests + no-tview green; zero
residual env-migration references in code or docs.
