+++ cmd/proxsave/main_footer.go
@@ -255,8 +255,6 @@ func printFinalSummaryCommands() {
 	fmt.Println("  --dry-run          - Test without changes")
 	fmt.Println("  --install          - Re-run interactive installation/setup")
 	fmt.Println("  --new-install      - Wipe installation directory (keep build/env/identity) then run installer")
-	fmt.Println("  --env-migration    - Run installer and migrate legacy Bash backup.env to Go template")
-	fmt.Println("  --env-migration-dry-run - Preview installer/migration without writing files")
 	fmt.Println("  --upgrade          - Update proxsave binary to latest release (also adds missing keys to backup.env)")
 	fmt.Println("  --newkey           - Generate a new encryption key for backups")
 	fmt.Println("  --decrypt          - Decrypt an existing backup archive")
