+++ docs/INSTALL.md
@@ -4,7 +4,6 @@
 
 - [🚀 Fast Install](#fast-install)
   - [Direct Install](#direct-install)
-  - [Migration](#migration)
   - [First Backup Workflow](#first-backup-workflow)
 - [⬆️ Upgrading ProxSave Binary](#upgrading-proxsave-binary)
   - [Quick Upgrade](#quick-upgrade)
@@ -14,9 +13,6 @@
   - [Prerequisites](#prerequisites)
   - [Building from Source](#building-from-source)
   - [Interactive Installation Wizard](#interactive-installation-wizard)
-- [🔄 Migrating an Existing Configuration](#migrating-an-existing-configuration-env-migration)
-  - [Migration Tools](#migration-tools)
-  - [Upgrade Steps](#upgrade-steps)
 
 ## Fast Install
 
@@ -48,20 +44,6 @@
 > `openssl` for this (preinstalled on Proxmox); the Go upgrade verifies it
 > natively. To verify a download yourself, see
 > [PROVENANCE_VERIFICATION.md](PROVENANCE_VERIFICATION.md#release-signature-sha256sumssig).
-
-### Migration
-
-1. Run migration installation from bash with old env file
-
-   ```bash
-   proxsave --env-migration
-   ```
-
-2. Run your first backup again after migration
-
-   ```bash
-   proxsave
-   ```
 
 ### First Backup Workflow
 
@@ -288,135 +270,3 @@ When shown, it does **not** modify your `backup.env`. It only:
 After completion, edit `configs/backup.env` manually for advanced options.
 
 `BASE_DIR` is detected from the installed `proxsave` executable. Do not add an active `BASE_DIR=...` line to `backup.env`; upgrades remove it and runtime ignores it if present.
-
----
-
-## Migrating an Existing Configuration (env-migration)
-
-The `--env-migration` tool imports an existing `backup.env` (e.g. from the legacy Bash version, v0.7.4-bash or earlier) into the current Go configuration, mapping most variables automatically. See [BACKUP_ENV_MAPPING.md](BACKUP_ENV_MAPPING.md) for the complete variable mapping.
-
-### Migration Tools
-
-#### Option 1: Interactive Tool
-
-Automatic tool based on variable mapping: BACKUP_ENV_MAPPING.md (we recommend checking after migration to ensure everything went smoothly)
-
-```bash
-proxsave --env-migration
-```
-
-You can then manually add your custom variables by referring to the mapping guide.
-
-#### Option 2: Migration Reference Guide
-
-The project includes a complete environment variable mapping guide to help you migrate your configuration:
-
-**[BACKUP_ENV_MAPPING.md](BACKUP_ENV_MAPPING.md)** - Complete Bash → Go variable mapping reference
-
-This guide categorizes every variable:
-
-- **SAME**: Variables with identical names (just copy them)
-- **RENAMED**: Variables with new names but automatic fallback (old names still work)
-- **SEMANTIC CHANGE**: Variables requiring value conversion (e.g., percentage → GB)
-- **LEGACY**: Bash-only variables no longer needed in Go
-
-**Quick migration workflow:**
-
-1. Open your Bash `backup.env`
-2. Open your Go `backup.env`
-3. Refer to `BACKUP_ENV_MAPPING.md` while copying your values
-4. Most variables can be copied directly (SAME + RENAMED categories)
-5. Pay attention to SEMANTIC CHANGE variables for manual conversion
-
-### Upgrade Steps
-
-1. **Run 1 minute Setup or Full manually Setup**
-
-2. **Migrate your configuration**
-
-   **Option A: Automatic migration (recommended for existing users)**
-
-   ```bash
-   # Step 1: Preview migration (recommended first step)
-   proxsave --env-migration-dry-run
-
-   # Review the output, then execute real migration
-   proxsave --env-migration
-
-   # The tool will:
-   # - Automatically map 70+ variables (SAME category)
-   # - Convert 16 renamed variables (RENAMED category)
-   # - Flag 2 variables for manual review (SEMANTIC CHANGE)
-   # - Skip 27 legacy variables (LEGACY category)
-   # - Create backup of existing config
-   ```
-
-   **Option B: Manual migration using mapping guide**
-
-   ```bash
-   # Edit with your Bash settings, using BACKUP_ENV_MAPPING.md as reference
-   nano configs/backup.env
-   ```
-
-3. **Test the configuration**
-
-   ```bash
-   # Dry-run to verify configuration
-   proxsave --dry-run
-
-   # Check the output for any warnings or errors
-   ```
-
-4. **Run a test backup**
-
-   ```bash
-   # First real backup
-   proxsave
-
-   # Verify results
-   ls -lh backup/
-   cat log/backup-*.log
-   ```
-
-### Key Migration Notes
-
-**Automatic variable fallbacks** - These old Bash variable names still work in Go:
-
-- `LOCAL_BACKUP_PATH` ? reads as `BACKUP_PATH`
-- `ENABLE_CLOUD_BACKUP` ? reads as `CLOUD_ENABLED`
-- `PROMETHEUS_ENABLED` ? reads as `METRICS_ENABLED`
-- And 13 more (see mapping guide for complete list)
-
-**Variables requiring conversion:**
-
-- `STORAGE_WARNING_THRESHOLD_PRIMARY="90"` (% used) ? `MIN_DISK_SPACE_PRIMARY_GB="1"` (GB free)
-- `CLOUD_BACKUP_PATH="/remote:path/folder"` (full path) ? `CLOUD_REMOTE="<remote>"` + `CLOUD_REMOTE_PATH="/folder"`
-
-**New Go-only features available:**
-
-- GFS retention policies (`RETENTION_POLICY=gfs`)
-- AGE encryption (`ENCRYPT_ARCHIVE=true`)
-- Parallel cloud uploads (`CLOUD_UPLOAD_MODE=parallel`)
-- Advanced security checks with auto-fix
-- Gotify and webhook notifications
-- Prometheus metrics export
-
-### Troubleshooting Migration
-
-**Problem**: "Configuration variable not recognized"
-
-- **Solution**: Check `BACKUP_ENV_MAPPING.md` to see if the variable was renamed or is now LEGACY
-
-**Problem**: Storage threshold warnings incorrect
-
-- **Solution**: Convert percentage-based thresholds to GB-based (SEMANTIC CHANGE variables)
-
-**Problem**: Cloud path not working
-
-- **Solution**: Split `CLOUD_BACKUP_PATH` into `CLOUD_REMOTE` (remote name) and `CLOUD_REMOTE_PATH` (full path inside that remote)
-
-**Still having issues?**
-
-- Review the complete mapping guide: [BACKUP_ENV_MAPPING.md](BACKUP_ENV_MAPPING.md)
-- Compare your Bash config with the Go template side-by-side
-- Enable debug logging: `proxsave --dry-run --log-level debug`
