+++ docs/README.md
@@ -23,10 +23,8 @@ below for the current operational and technical behavior.
 
 ## Supporting References
 
-- [BACKUP_ENV_MAPPING.md](BACKUP_ENV_MAPPING.md): legacy Bash to Go env mapping
 - [CLOUD_STORAGE.md](CLOUD_STORAGE.md): cloud/rclone behavior
 - [ENCRYPTION.md](ENCRYPTION.md): archive encryption and decrypt/restore flow
 - [PROVENANCE_VERIFICATION.md](PROVENANCE_VERIFICATION.md): attestation verification
-- [MIGRATION_GUIDE.md](MIGRATION_GUIDE.md): env-migration of an existing backup.env to Go config
 - [CLUSTER_RECOVERY.md](CLUSTER_RECOVERY.md): PVE cluster disaster recovery
 - [RELEASE-PROCESS.md](RELEASE-PROCESS.md): release engineering notes
