+++ internal/cli/args.go
@@ -39,10 +39,7 @@ type Args struct {
 	UpgradeConfig     bool
 	UpgradeConfigDry  bool
 	UpgradeConfigJSON bool
-	EnvMigration      bool
-	EnvMigrationDry   bool
 	CleanupGuards     bool
-	LegacyEnvPath     string
 	Backup            bool
 	Daemon            bool
 	DaemonSetup       bool
@@ -113,14 +110,8 @@ func Parse() *Args {
 		"Reset the installation directory (preserving build/env/identity) and launch the interactive installer")
 	flag.BoolVar(&args.Upgrade, "upgrade", false,
 		"Download and install the latest ProxSave binary (also upgrades backup.env by adding missing keys from the new template). Append 'y' to auto-confirm (e.g., --upgrade y)")
-	flag.BoolVar(&args.EnvMigration, "env-migration", false,
-		"Run the installer and migrate a legacy Bash backup.env to the Go template")
-	flag.BoolVar(&args.EnvMigrationDry, "env-migration-dry-run", false,
-		"Preview the installer + legacy env migration without writing files")
 	flag.BoolVar(&args.CleanupGuards, "cleanup-guards", false,
 		"Cleanup ProxSave guard bind mounts and directories (/var/lib/proxsave/guards). Use with --dry-run to preview")
-	flag.StringVar(&args.LegacyEnvPath, "old-env", "",
-		"Path to the legacy Bash backup.env used during --env-migration")
 
 	flag.BoolVar(&args.UpgradeConfig, "upgrade-config", false,
 		"Upgrade configuration file using the embedded template (adds missing keys, preserves existing and custom keys)")
